暂时关闭Security支持 调整ViewController
