+++ EndlessMine-WebServer/src/main/java/keyto/endlessmine/webserver/controller/LoginController.java
@@ -16,48 +16,20 @@
  *
  * E-mail: keyto1995@outlook.com
  */
-package keyto.endlessmine.webserver.eneity;
+package keyto.endlessmine.webserver.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 /**
  *
  * @author Keyto
  */
-public class User {
-
-    private String name;
-    private String password;
-
-    /**
-     * @return the name
-     */
-    public String getName() {
-        return name;
+@Controller
+public class LoginController {
+    @RequestMapping("do_login")
+    String do_login(){
+        
+        return "success";
     }
-
-    /**
-     * @param name the name to set
-     */
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    /**
-     * @return the password
-     */
-    public String getPassword() {
-        return password;
-    }
-
-    /**
-     * @param password the password to set
-     */
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    @Override
-    public String toString() {
-        return super.toString() + " [ " + name + "," + password + " ]";
-    }
-
 }
