+++ EndlessMine-WebServer/src/main/java/keyto/endlessmine/webserver/controller/QuickViewController.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2017 Keyto
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * E-mail: keyto1995@outlook.com
+ */
+package keyto.endlessmine.webserver.controller;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+/**
+ *
+ * @author Keyto
+ */
+@Configuration
+public class QuickViewController extends WebMvcConfigurerAdapter{
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/index").setViewName("index");
+        registry.addViewController("/signUpPage").setViewName("signUpPage");
+        registry.addViewController("/loginPage").setViewName("loginPage");
+        registry.addViewController("/gamePage").setViewName("gamePage");
+        
+    }
+    
+}
