+++ EndlessMine-WebServer/src/main/java/keyto/endlessmine/webserver/massage/MsgDoLogin.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright (C) 2017 Keyto
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * E-mail: keyto1995@outlook.com
+ */
+package keyto.endlessmine.webserver.massage;
+
+/**
+ *
+ * @author Keyto
+ */
+public class MsgDoLogin {
+
+    private String name;
+    private String password;
+
+    /**
+     * @return the name
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * @param name the name to set
+     */
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    /**
+     * @return the password
+     */
+    public String getPassword() {
+        return password;
+    }
+
+    /**
+     * @param password the password to set
+     */
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    @Override
+    public String toString() {
+        return super.toString()
+                + " [ name=" + name
+                + ",password=" + password 
+                + " ]";
+    }
+
+}
