--- EndlessMine-WebServer/src/main/resources/templates/showPage.html
@@ -1,91 +1,0 @@
-<!DOCTYPE html>
-<!--
-Copyright (C) 2017 Keyto
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-E-mail: keyto1995@outlook.com
--->
-<html lang="zh-CN" 
-      xmlns:th="http://www.thymeleaf.org">
-    <head>
-        <meta charset="UTF-8"/>
-        <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-        <meta name="viewport" content="width=device-width, initial-scale=1"/>
-        <!-- 上述3个meta标签*必须*放在最前面，任何其他内容都*必须*跟随其后！ -->
-        <title>TODO supply a title</title>
-        <!-- Bootstrap -->
-        <link href="../static/bootstrap/css/bootstrap.min.css" th:href="@{bootstrap/css/bootstrap.min.css}" rel="stylesheet"/>
-
-        <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
-        <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-        <!--[if lt IE 9]>
-          <script src="https://cdn.bootcss.com/html5shiv/3.7.3/html5shiv.min.js"></script>
-          <script src="https://cdn.bootcss.com/respond.js/1.4.2/respond.min.js"></script>
-        <![endif]-->
-    </head>
-    <body>
-        <button id="ca">click here</button>
-        <div>TODO write content</div>
-        <button id="chunk"> get a chunk </button>
-
-
-        <!-- jQuery (necessary for Bootstrap's JavaScript plugins) -->
-        <!--<script src="//cdn.bootcss.com/jquery/3.2.1/jquery.min.js"></script>-->
-        <script src="../static/jquery-3.2.1.min.js" th:src="@{jquery-3.2.1.min.js}"></script>
-        <!-- Include all compiled plugins (below), or include individual files as needed -->
-        <script src="../static/bootstrap/js/bootstrap.min.js" th:src="@{bootstrap/js/bootstrap.min.js}"></script>
-        <script src="../static/script/EndlessMineObj.js" th:src="@{script/EndlessMineObj.js}"></script>
-        <script type="text/javascript">
-
-
-
-            $("#ca").click(function () {
-                $.post("demo_test_show",
-                        {
-                            chunkPointX: 1,
-                            chunkPointY: 2,
-                            blockIndex: 3,
-                            mouseButton: "PRIMARY"
-                        },
-                        function (data, status) {
-                            for (var o in data) {
-                                var pic = new Block(data[o]).blockInfo.getPic();
-                                console.log("Pic:" + pic);
-                            }
-                            console.log(data);
-                        });
-            });
-
-            $("#chunk").click(function () {
-                $.getJSON("demo_test_show_chunk",
-                        {
-                            chunkPointX: 1,
-                            chunkPointY: 2
-                        },
-                        function (data, status) {
-                            for (var o in data) {
-                                var pic = new BlockInfo(data[o]).getPic();
-                                console.log("Pic:" + o + "->" + pic);
-                            }
-//                            alert("Data: " + data + "\nStatus: " + status);
-                        });
-            });
-        </script>
-    </body>
-</html>
-
-
-
-
